requirements.txt
flask>=2.2.2
flask_restful>=0.3.8
flask_sqlalchemy>=2.0
flask_migrate>=2.7.0
flask_marshmallow>=0.14.0
marshmallow-sqlalchemy>=0.24.2
psycopg2-binary>=2.8.3
SQLAlchemy>=1.3.7
python-decouple